requirements.txt
aioredis==1.2.0
amqp==2.3.2
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.1
attrs==18.1.0
autobahn==18.7.1
Automat==0.7.0
autopep8==1.4.2
backcall==0.1.0
beautifulsoup4==4.6.1
billiard==3.5.0.4
binance==0.3
cachetools==2.1.0
celery==4.2.1
certifi==2018.8.13
cffi==1.11.5
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
coverage==4.5.2
cryptography==2.3
cycler==0.10.0
daphne==2.2.5
dateparser==0.7.0
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1.1
django-appconf==1.0.2
django-bootstrap4==0.0.7
django-compressor==2.2
django-cors-headers==2.5.2
django-countries==5.3.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-extensions==2.1.5
django-heroku==0.3.1
django-jet==1.0.8
djangorestframework==3.9.1
djangorestframework-simplejwt==4.1.0
Faker==1.0.2
future==0.16.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
gspread==3.0.1
gunicorn==19.7.0
hiredis==1.0.0
httplib2==0.11.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
kiwisolver==1.0.1
kombu==4.2.1
lazy-object-proxy==1.3.1
matplotlib==2.2.3
mccabe==0.6.1
mpl-finance==0.10.0
msgpack==0.6.1
numpy==1.15.0
oauth2client==4.1.2
pandas==0.23.4
parso==0.3.1
pbr==4.2.0
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
proxyscrape==0.1.0
psutil==5.4.7
psycopg2==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.18
Pygments==2.2.0
pygsheets==1.1.4
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.1.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
python-binance==0.7.0
python-dateutil==2.7.3
python-telegram-bot==10.1.0
pytz==2018.5
rcssmin==1.0.6
redis==2.10.6
regex==2018.7.11
requests==2.19.1
rjsmin==1.0.12
rsa==3.4.2
selenium==3.14.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
stevedore==1.29.0
text-unidecode==1.2
traitlets==4.3.2
Twisted==18.7.0
txaio==18.7.1
typed-ast==1.1.0
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wcwidth==0.1.7
whitenoise==4.1.2
wrapt==1.10.11
zope.interface==4.5.0
